+++ jpa/eclipselink.jpa.testapps/jpa.test.batchfetch/src/main/java/org/eclipse/persistence/testing/models/jpa/batchfetch/Company.java
@@ -12,7 +12,7 @@
 
 // Contributors:
 
-package org.eclipse.persistence.jpa.testapps.batchfetch;
+package org.eclipse.persistence.testing.models.jpa.batchfetch;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
